# openapi/personalcrm_openapi_v1.yaml
openapi: 3.0.3
info:
  title: PersonalCRM API
  version: 1.0.0
  description: |
    PersonalCRM v1 API for managing contacts, interactions, introductions, reminders,
    strategy/wheel, projects, and RBAC in a cloud workspace with assistant access.

    Notes:
    - All requests are scoped to a workspace.
    - Field-level privacy is enforced server-side; assistants may receive redacted fields.
    - Use cursor-based pagination (next_cursor).
servers:
  - url: https://api.personalcrm.example.com
    description: Production (replace)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Search by name
      responses:
        '200':
          description: Organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrganizations'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Organizations]
      summary: Create an organization
      operationId: createOrganization
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      operationId: listContacts
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search across name, org, title, shared notes
        - name: tags
          in: query
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
          description: Filter by tags (OR semantics)
        - name: tier
          in: query
          schema:
            $ref: '#/components/schemas/TieStrength'
          description: Filter by tie strength (close/medium/weak)
        - name: visibility
          in: query
          schema:
            $ref: '#/components/schemas/Visibility'
          description: Filter by record visibility
        - name: updated_since
          in: query
          schema:
            type: string
            format: date-time
          description: Return contacts updated after this timestamp
        - name: sort
          in: query
          schema:
            type: string
            enum: [updated_desc, last_interaction_desc, name_asc]
          description: Sort option
      responses:
        '200':
          description: Contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContacts'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Contacts]
      summary: Create a contact
      operationId: createContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/contacts/{contact_id}:
    get:
      tags: [Contacts]
      summary: Get a contact
      operationId: getContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      responses:
        '200':
          description: Contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Contacts]
      summary: Update a contact
      operationId: updateContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Contacts]
      summary: Delete a contact
      operationId: deleteContact
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/contacts/{contact_id}/interactions:
    get:
      tags: [Interactions]
      summary: List interactions for a contact
      operationId: listContactInteractions
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/InteractionType'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Interactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInteractions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Interactions]
      summary: Create an interaction for a contact
      operationId: createContactInteraction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ContactIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/interactions/{interaction_id}:
    get:
      tags: [Interactions]
      summary: Get an interaction
      operationId: getInteraction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/InteractionIdPath'
      responses:
        '200':
          description: Interaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Interactions]
      summary: Update an interaction
      operationId: updateInteraction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/InteractionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Interactions]
      summary: Delete an interaction
      operationId: deleteInteraction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/InteractionIdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/introductions:
    get:
      tags: [Introductions]
      summary: List introductions
      operationId: listIntroductions
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IntroductionStatus'
        - name: contact_id
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
          description: Filter by a participant contact id
      responses:
        '200':
          description: Introductions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIntroductions'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Introductions]
      summary: Create an introduction workflow
      operationId: createIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/introductions/{introduction_id}:
    get:
      tags: [Introductions]
      summary: Get an introduction
      operationId: getIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/IntroductionIdPath'
      responses:
        '200':
          description: Introduction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Introductions]
      summary: Update an introduction
      operationId: updateIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/IntroductionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceIdQuery:
      name: workspace_id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
      description: Workspace scope for the request
...
    ContactIdPath:
      name: contact_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
    InteractionIdPath:
      name: interaction_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UUID:
      type: string
      format: uuid
...
    Visibility:
      type: string
      enum: [private, shared, limited]
      description: |
        Record visibility.
        - private: only Owner (and roles with private access)
        - shared: Owner + Assistant
        - limited: Assistant can see a redacted subset
    TieStrength:
      type: string
      enum: [close, medium, weak]
    InteractionType:
      type: string
      enum: [meeting, call, message, event, intro, help_given, help_received, note]
...
    IntroductionStatus:
      type: string
      enum: [requested, approved_a, approved_b, sent, met, completed, canceled]
...
    Organization:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        website: { type: string }
        industries:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, name, created_at, updated_at]
    OrganizationCreate:
      type: object
      properties:
        name: { type: string }
        website: { type: string }
        industries:
          type: array
          items: { type: string }
      required: [name]
    Contact:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        visibility: { $ref: '#/components/schemas/Visibility' }

        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }

        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
          description: Map of messenger name -> handle

        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }

        organization:
          nullable: true
          $ref: '#/components/schemas/Organization'
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }

        tie_strength: { $ref: '#/components/schemas/TieStrength' }

        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
          description: May be redacted for assistants depending on privacy
        emotional_balance:
          type: number
          format: double
          nullable: true
          description: May be redacted for assistants depending on privacy

        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }

        met_context:
          type: string
          description: Where/how you met
        shared_notes:
          type: string
        private_notes:
          type: string
          description: Owner-only notes (may be null/redacted)

        last_interaction_at: { type: string, format: date-time, nullable: true }
        next_touch_at: { type: string, format: date-time, nullable: true }

        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        - id
        - workspace_id
        - visibility
        - display_name
        - tie_strength
        - created_at
        - updated_at
    ContactCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }
        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }
        organization_id: { $ref: '#/components/schemas/UUID' }
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        tie_strength: { $ref: '#/components/schemas/TieStrength' }
        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        emotional_balance:
          type: number
          format: double
          nullable: true
        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }
        met_context: { type: string }
        shared_notes: { type: string }
        private_notes: { type: string }
        next_touch_at: { type: string, format: date-time }
      required: [display_name, tie_strength]
    ContactUpdate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        display_name: { type: string }
        photo_url: { type: string }
        emails:
          type: array
          items: { type: string, format: email }
        phones:
          type: array
          items: { type: string }
        messengers:
          type: object
          additionalProperties: { type: string }
        city: { type: string }
        timezone: { type: string }
        birthday: { type: string, format: date }
        organization_id: { $ref: '#/components/schemas/UUID' }
        job_title: { type: string }
        industries:
          type: array
          items: { type: string }
        competencies:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        tie_strength: { $ref: '#/components/schemas/TieStrength' }
        trust_score:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        emotional_balance:
          type: number
          format: double
          nullable: true
        how_can_help:
          type: array
          items: { type: string }
        how_i_can_help:
          type: array
          items: { type: string }
        met_context: { type: string }
        shared_notes: { type: string }
        private_notes: { type: string }
        next_touch_at: { type: string, format: date-time }
    Interaction:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        contact_id: { $ref: '#/components/schemas/UUID' }
        visibility: { $ref: '#/components/schemas/Visibility' }
        type: { $ref: '#/components/schemas/InteractionType' }
        channel: { $ref: '#/components/schemas/InteractionChannel' }
        occurred_at: { type: string, format: date-time }
        summary: { type: string }
        outcome: { type: string }
        next_action: { type: string }
        next_action_at: { type: string, format: date-time, nullable: true }
        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, contact_id, type, occurred_at, created_at, updated_at]
    InteractionCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        type: { $ref: '#/components/schemas/InteractionType' }
        channel: { $ref: '#/components/schemas/InteractionChannel' }
        occurred_at: { type: string, format: date-time }
        summary: { type: string }
        outcome: { type: string }
        next_action: { type: string }
        next_action_at: { type: string, format: date-time }
      required: [type, occurred_at]
    InteractionUpdate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        type: { $ref: '#/components/schemas/InteractionType' }
        channel: { $ref: '#/components/schemas/InteractionChannel' }
        occurred_at: { type: string, format: date-time }
        summary: { type: string }
        outcome: { type: string }
        next_action: { type: string }
        next_action_at: { type: string, format: date-time, nullable: true }
    Introduction:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        status: { $ref: '#/components/schemas/IntroductionStatus' }

        requester_contact_id: { $ref: '#/components/schemas/UUID' }
        introducer_contact_id: { $ref: '#/components/schemas/UUID' }
        target_contact_id: { $ref: '#/components/schemas/UUID' }

        ask: { type: string }
        benefit_for_requester: { type: string }
        benefit_for_target: { type: string }

        consent_requester: { type: boolean }
        consent_target: { type: boolean }

        sent_at: { type: string, format: date-time, nullable: true }
        met_at: { type: string, format: date-time, nullable: true }
        outcome: { type: string, nullable: true }

        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        - id
        - workspace_id
        - status
        - requester_contact_id
        - introducer_contact_id
        - target_contact_id
        - ask
        - created_at
        - updated_at
    IntroductionCreate:
      type: object
      properties:
        requester_contact_id: { $ref: '#/components/schemas/UUID' }
        introducer_contact_id: { $ref: '#/components/schemas/UUID' }
        target_contact_id: { $ref: '#/components/schemas/UUID' }
        ask: { type: string }
        benefit_for_requester: { type: string }
        benefit_for_target: { type: string }
        consent_requester: { type: boolean, default: false }
        consent_target: { type: boolean, default: false }
        status: { $ref: '#/components/schemas/IntroductionStatus' }
      required: [requester_contact_id, introducer_contact_id, target_contact_id, ask]
...
    PaginatedContacts:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Contact' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
    PaginatedOrganizations:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
    PaginatedInteractions:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Interaction' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
    PaginatedIntroductions:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Introduction' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...